+++ crates/holochain_conductor_api/CHANGELOG.md
@@ -4,6 +4,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+- BREAKING: CONDUCTOR CONFIG CHANGE--related to update to lair 0.0.3
+  - `passphrase_service` is now required
+    - The only implemented option is `danger_insecure_from_config`
+
+#### Example
+
+```yaml
+---
+passphrase_service:
+  type: danger_insecure_from_config
+  passphrase: "foobar"
+```
+
 ## 0.0.2
 
 ## 0.0.1